+++ app/scripts/google.js
@@ -7,15 +7,13 @@
   head.appendChild(script);
 })();
 
-window._gaq = window._gaq || [];
-window._gaq.push(['_setAccount', 'UA-60316581-1']);
-window._gaq.push(['_trackPageview']);
+var _gaq = _gaq || [];
+_gaq.push(['_setAccount', 'UA-60316581-1']);
+_gaq.push(['_setCustomVar', 1, 'DIMVersion', '$DIM_VERSION', 3]);
+_gaq.push(['_trackPageview']);
 
-(function loadGA() {
-  var ga = document.createElement('script');
-  ga.type = 'text/javascript';
-  ga.async = true;
-  ga.src = 'https://ssl.google-analytics.com/ga.js';
-  var s = document.getElementsByTagName('script')[0];
-  s.parentNode.insertBefore(ga, s);
+(function() {
+  var ga = document.createElement('script'); ga.type = 'text/javascript'; ga.async = true;
+  ga.src = 'https://www.google-analytics.com/ga.js';
+  var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(ga, s);
 })();