+++ app/scripts/minmax/dimMinMax.controller.js
@@ -9,6 +9,7 @@
   function dimMinMaxCtrl($scope, $state, $q, $timeout, $location, loadingTracker, dimStoreService, dimItemService, ngDialog) {
     var vm = this;
     var buckets = [];
+    var vendorBuckets = [];
 
     function getBonusType(armorpiece) {
       if (!armorpiece.normalStats) {
@@ -38,7 +39,7 @@
       };
     }
 
-    function getBestArmor(bucket, locked) {
+    function getBestArmor(bucket, vendorBucket, locked, excluded) {
       var statHashes = [
           { stats: [144602215, 1735777505], type: 'intdisc' },
           { stats: [144602215, 4244567218], type: 'intstr' },
@@ -52,20 +53,28 @@
       var curbest;
       var bestCombs;
       var armortype;
+
       for (armortype in bucket) {
+        var combined = (vm.includeVendors) ? bucket[armortype].concat(vendorBucket[armortype]) : bucket[armortype];
         if (locked[armortype]) {
           best = [{ item: locked[armortype], bonus_type: getBonusType(locked[armortype]) }];
         } else {
           best = [];
+
+          // Filter out excluded
+          var filtered = _.filter(combined, function(item) {
+            return !_.findWhere(excluded, { index: item.index });
+          });
           statHashes.forEach(function(hash, index) {
             if (!vm.mode && index > 2) {
               return;
             }
-            curbest = getBestItem(bucket[armortype], hash.stats, hash.type);
+
+            curbest = getBestItem(filtered, hash.stats, hash.type);
             best.push(curbest);
             // add the best -> if best is exotic -> get best legendary
             if (curbest.item.isExotic && armortype !== 'ClassItem') {
-              best.push(getBestItem(bucket[armortype], hash.stats, hash.type, true));
+              best.push(getBestItem(filtered, hash.stats, hash.type, true));
             }
           });
         }
@@ -99,6 +108,27 @@
       ) < 2;
     }
 
+    function getId(index) {
+      var split = index.split('-');
+      return split[1] === '1' ? index : split[1];
+    }
+
+    function getItemById(id, type) {
+      return _.findWhere(buckets[vm.active][type], { id: id }) || _.findWhere(vendorBuckets[vm.active][type], { index: id });
+    }
+
+    function alreadyExists(set, id) {
+      return _.findWhere(set, { id: id }) || _.findWhere(set, { index: id });
+    }
+
+    function mergeBuckets(bucket1, bucket2) {
+      var merged = {};
+      _.each(_.keys(bucket1), function(type) {
+        merged[type] = bucket1[type].concat(bucket2[type]);
+      });
+      return merged;
+    }
+
     angular.extend(vm, {
       active: 'warlock',
       activesets: '5/5/1',
@@ -107,8 +137,11 @@
       scale_type: 'scaled',
       allSetTiers: [],
       highestsets: {},
+      excludeditems: [],
       lockeditems: { Helmet: null, Gauntlets: null, Chest: null, Leg: null, ClassItem: null, Artifact: null, Ghost: null },
       type: 'Helmet',
+      vendorType: 'Helmet',
+      includeVendors: false,
       showBlues: false,
       showExotics: true,
       showYear1: false,
@@ -117,8 +150,12 @@
       statOrder: '-stats.STAT_INTELLECT.value',
       ranked: {},
       lockedItemsValid: function(droppedId, droppedType) {
-        droppedId = droppedId.split('-')[1];
-        var item = _.findWhere(buckets[vm.active][droppedType], { id: droppedId });
+        droppedId = getId(droppedId);
+        if (alreadyExists(vm.excludeditems, droppedId)) {
+          return false;
+        }
+
+        var item = getItemById(droppedId, droppedType);
         var exoticCount = ((item.isExotic && item.type !== 'ClassItem') ? 1 : 0);
         _.each(vm.lockeditems, function(lockeditem) {
           if (lockeditem === null || lockeditem.type === droppedType) {
@@ -130,24 +167,46 @@
         });
         return exoticCount < 2;
       },
-
+      excludedItemsValid: function(droppedId, droppedType) {
+        return !(vm.lockeditems[droppedType] && alreadyExists([vm.lockeditems[droppedType]], droppedId));
+      },
       onCharacterChange: function() {
-        vm.ranked = buckets[vm.active];
+        vm.ranked = (vm.includeVendors) ? mergeBuckets(buckets[vm.active], vendorBuckets[vm.active]) : buckets[vm.active];
         vm.lockeditems = { Helmet: null, Gauntlets: null, Chest: null, Leg: null, ClassItem: null, Artifact: null, Ghost: null };
+        vm.excludeditems = [];
         vm.highestsets = vm.getSetBucketsStep(vm.active);
       },
       onModeChange: function() {
+        vm.highestsets = vm.getSetBucketsStep(vm.active);
+      },
+      onIncludeVendorsChange: function() {
+        if (vm.includeVendors) {
+          vm.ranked = mergeBuckets(buckets[vm.active], vendorBuckets[vm.active]);
+        } else {
+          vm.ranked = buckets[vm.active];
+
+          // Filter any vendor items from locked or excluded items
+          _.each(vm.lockeditems, function(item, type) {
+            if (item && item.isVendorItem) {
+              vm.lockeditems[type] = null;
+            }
+          });
+
+          vm.excludeditems = _.filter(vm.excludeditems, function(item) {
+            return !item.isVendorItem;
+          });
+        }
         vm.highestsets = vm.getSetBucketsStep(vm.active);
       },
       onOrderChange: function() {
         vm.setOrderValues = vm.setOrder.split(',');
       },
       onDrop: function(droppedId, type) {
-        droppedId = droppedId.split('-')[1];
-        if (vm.lockeditems[type] && vm.lockeditems[type].id === droppedId) {
+        droppedId = getId(droppedId);
+        if (vm.lockeditems[type] && alreadyExists([vm.lockeditems[type]], droppedId)) {
           return;
         }
-        var item = _.findWhere(buckets[vm.active][type], { id: droppedId });
+        var item = getItemById(droppedId, type);
         vm.lockeditems[type] = item;
         vm.highestsets = vm.getSetBucketsStep(vm.active);
         if (vm.progress < 1.0) {
@@ -160,6 +219,29 @@
         vm.highestsets = vm.getSetBucketsStep(vm.active);
         if (vm.progress < 1.0) {
           vm.lockedchanged = true;
+        }
+      },
+      excludeItem: function(item) {
+        vm.onExcludedDrop(item.index, item.type);
+      },
+      onExcludedDrop: function(droppedId, type) {
+        droppedId = getId(droppedId);
+        if (alreadyExists(vm.excludeditems, droppedId)) {
+          return;
+        }
+        var item = getItemById(droppedId, type);
+        vm.excludeditems.push(item);
+        vm.highestsets = vm.getSetBucketsStep(vm.active);
+        if (vm.progress < 1.0) {
+          vm.excludedchanged = true;
+        }
+      },
+      onExcludedRemove: function(removedIndex) {
+        vm.excludeditems = _.filter(vm.excludeditems, function(excludeditem) { return excludeditem.index !== removedIndex; });
+
+        vm.highestsets = vm.getSetBucketsStep(vm.active);
+        if (vm.progress < 1.0) {
+          vm.excludedchanged = true;
         }
       },
       newLoadout: function(set) {
@@ -181,7 +263,7 @@
         });
       },
       getSetBucketsStep: function(activeGaurdian) {
-        var bestArmor = getBestArmor(buckets[activeGaurdian], vm.lockeditems);
+        var bestArmor = getBestArmor(buckets[activeGaurdian], vendorBuckets[activeGaurdian], vm.lockeditems, vm.excludeditems);
         var helms = bestArmor.Helmet || [];
         var gaunts = bestArmor.Gauntlets || [];
         var chests = bestArmor.Chest || [];
@@ -264,9 +346,10 @@
 
                         processedCount++;
                         if (processedCount % 50000 === 0) { // do this so the page doesn't lock up
-                          if (vm.active !== activeGaurdian || vm.lockedchanged || $location.path() !== '/best') {
+                          if (vm.active !== activeGaurdian || vm.lockedchanged || vm.excludedchanged || $location.path() !== '/best') {
                             // If active gaurdian or page is changed then stop processing combinations
                             vm.lockedchanged = false;
+                            vm.excludedchanged = false;
                             return;
                           }
                           vm.progress = processedCount / combos;
@@ -303,6 +386,7 @@
         }
         console.time('elapsed');
         vm.lockedchanged = false;
+        vm.excludedchanged = false;
         $timeout(step, 0, true, activeGaurdian, 0, 0, 0, 0, 0, 0, 0, 0);
         return setMap;
       },
@@ -318,7 +402,8 @@
         }
 
         var allItems = [];
-        vm.active = _.sortBy(stores, 'lastPlayed').reverse()[0].class.toLowerCase() || 'warlock';
+        var vendorItems = [];
+        vm.active = dimStoreService.getActiveStore().class.toLowerCase() || 'warlock';
 
         _.each(stores, function(store) {
           var items = _.filter(store.items, function(item) {
@@ -330,6 +415,18 @@
           });
 
           allItems = allItems.concat(items);
+
+          _.each(store.vendors, function(vendor) {
+            var vendItems = _.filter(vendor.items.armor, function(item) {
+              return item.primStat &&
+              item.primStat.statHash === 3897883278 && // has defense hash
+              ((vm.showBlues && item.tier === 'Rare') || item.tier === 'Legendary' || (vm.showExotics && item.isExotic)) && // is legendary or exotic
+              item.primStat.value >= 280 && // only 280+ light items
+              item.stats;
+            });
+
+            vendorItems = vendorItems.concat(vendItems);
+          });
         });
 
         function normalizeStats(item) {
@@ -375,8 +472,9 @@
         }
 
         function initBuckets() {
-          function loadBucket(classType) {
-            return getBuckets(allItems.filter(function(item) {
+          function loadBucket(classType, useVendorItems = false) {
+            var items = (useVendorItems) ? vendorItems : allItems;
+            return getBuckets(items.filter(function(item) {
               return item.classType === classType || item.classType === 3;
             }));
           }
@@ -384,6 +482,12 @@
             titan: loadBucket(0),
             hunter: loadBucket(1),
             warlock: loadBucket(2)
+          };
+
+          vendorBuckets = {
+            titan: loadBucket(0, true),
+            hunter: loadBucket(1, true),
+            warlock: loadBucket(2, true)
           };
         }