+++ app/scripts/services/dimSettingsService.factory.js
@@ -26,6 +26,8 @@
       itemDetails: false,
       // Show item quality percentages
       itemQuality: false,
+      // Show new items with an overlay
+      showNewItems: false,
       // Show animation of new item overlay on new items
       showNewAnimation: true,
       // Show elemental damage icons
@@ -41,6 +43,8 @@
       vaultMaxCol: 999,
       // How big in pixels to draw items
       itemSize: 44,
+      // Which categories or buckets should be collapsed?
+      collapsedSections: {},
 
       save: function() {
         if (!_loaded) {