+++ app/scripts/shell/dimSearchFilter.directive.js
@@ -31,7 +31,7 @@
   var filterTrans = {
     dmg: ['arc', 'solar', 'void', 'kinetic'],
     type: ['primary', 'special', 'heavy', 'helmet', 'leg', 'gauntlets', 'chest', 'class', 'classitem', 'artifact', 'ghost', 'horn', 'consumable', 'ship', 'material', 'vehicle', 'emblem', 'bounties', 'quests', 'messages', 'missions', 'emote'],
-    tier: ['common', 'uncommon', 'rare', 'legendary', 'exotic'],
+    tier: ['common', 'uncommon', 'rare', 'legendary', 'exotic', 'white', 'green', 'blue', 'purple', 'yellow'],
     incomplete: ['incomplete'],
     complete: ['complete'],
     xpcomplete: ['xpcomplete'],
@@ -52,12 +52,13 @@
     year: ['year1', 'year2'],
     infusable: ['infusable', 'infuse'],
     stattype: ['intellect', 'discipline', 'strength'],
-    new: ['new']
+    new: ['new'],
+    glimmer: ['glimmeritem', 'glimmerboost', 'glimmersupply']
   };
 
-  var keywords = _.flatten(_.values(filterTrans)).map(function(word) {
-    return "is:" + word;
-  });
+  var keywords = _.flatten(_.flatten(_.values(filterTrans)).map(function(word) {
+    return ["is:" + word, "not:" + word];
+  }));
 
   // Filters that operate on ranges (>, <, >=, <=)
   var ranges = ['light', 'level', 'quality', 'percentage'];
@@ -69,7 +70,7 @@
     element.find('input').textcomplete([
       {
         words: keywords,
-        match: /\b((li|le|qu|pe|is:)\w*)$/,
+        match: /\b((li|le|qu|pe|is:|not:)\w*)$/,
         search: function(term, callback) {
           callback($.map(this.words, function(word) {
             return word.indexOf(term) === 0 ? word : null;
@@ -77,7 +78,8 @@
         },
         index: 1,
         replace: function(word) {
-          return word.indexOf('is:') === 0 ? (word + ' ') : word;
+          return (word.indexOf('is:') === 0 && word.indexOf('not:') === 0)
+            ? (word + ' ') : word;
         }
       }
     ], {
@@ -85,9 +87,9 @@
     });
   }
 
-  SearchFilterCtrl.$inject = ['$scope', 'dimStoreService', 'dimVendorService', 'dimSearchService'];
+  SearchFilterCtrl.$inject = ['$scope', 'dimStoreService', 'dimSearchService'];
 
-  function SearchFilterCtrl($scope, dimStoreService, dimVendorService, dimSearchService) {
+  function SearchFilterCtrl($scope, dimStoreService, dimSearchService) {
     var vm = this;
     var filterInputSelector = '#filter-input';
     var _duplicates = null; // Holds a map from item hash to count of occurrances of that hash
@@ -150,8 +152,8 @@
       var filterFn;
       var filters = [];
 
-      function addPredicate(predicate, filter) {
-        filters.push({ predicate: predicate, value: filter });
+      function addPredicate(predicate, filter, invert = false) {
+        filters.push({ predicate: predicate, value: filter, invert: invert });
       }
 
       _.each(searchTerms, function(term) {
@@ -161,11 +163,26 @@
             predicate = _cachedFilters[filter];
             addPredicate(predicate, filter);
           } else {
-            for (var key in filterTrans) {
+            for (const key in filterTrans) {
               if (filterTrans.hasOwnProperty(key) && filterTrans[key].indexOf(filter) > -1) {
                 predicate = key;
                 _cachedFilters[filter] = key;
                 addPredicate(predicate, filter);
+                break;
+              }
+            }
+          }
+        } else if (term.indexOf('not:') >= 0) {
+          filter = term.replace('not:', '');
+          if (_cachedFilters[filter]) {
+            predicate = _cachedFilters[filter];
+            addPredicate(predicate, filter, true);
+          } else {
+            for (const key in filterTrans) {
+              if (filterTrans.hasOwnProperty(key) && filterTrans[key].indexOf(filter) > -1) {
+                predicate = key;
+                _cachedFilters[filter] = key;
+                addPredicate(predicate, filter, true);
                 break;
               }
             }
@@ -183,7 +200,8 @@
 
       filterFn = function(item) {
         return _.all(filters, function(filter) {
-          return filterFns[filter.predicate](filter.value, item);
+          var result = filterFns[filter.predicate](filter.value, item);
+          return filter.invert ? !result : result;
         });
       };
 
@@ -191,37 +209,17 @@
         _.each(store.items, function(item) {
           item.visible = (filters.length > 0) ? filterFn(item) : true;
         });
-      });
 
-      if (dimVendorService.vendorItems) {
-        var setVisible = function(vendor) {
-          _.each(vendor.items, function(classType) {
-            _.each(classType, function(armorType) {
-              _.each(armorType, function(item) {
-                item.visible = (filters.length > 0) ? filterFn(item) : true;
-              });
-            });
+        // Filter vendor items
+        _.each(store.vendors, function(vendor) {
+          _.each(vendor.items.armor, function(item) {
+            item.visible = (filters.length > 0) ? filterFn(item) : true;
           });
-        };
-
-        setVisible(dimVendorService.vendorItems.crucible.Crucible);
-        setVisible(dimVendorService.vendorItems.exotics.Exotics);
-        _.each(dimVendorService.vendorItems.factions, function(faction) {
-          setVisible(faction);
+          _.each(vendor.items.weapons, function(item) {
+            item.visible = (filters.length > 0) ? filterFn(item) : true;
+          });
         });
-        _.each(dimVendorService.vendorItems.misc, function(miscVendor) {
-          setVisible(miscVendor);
-        });
-        _.each(dimVendorService.vendorItems.misc, function(miscVendor) {
-          setVisible(miscVendor);
-        });
-        _.each(dimVendorService.vendorItems.vanguard, function(vanguardVendor) {
-          setVisible(vanguardVendor);
-        });
-        if (dimVendorService.vendorItems.banner) {
-          setVisible(dimVendorService.vendorItems.banner.Banner);
-        }
-      }
+      });
     };
 
     // Cache for searches against filterTrans. Somewhat noticebly speeds up the lookup on my older Mac, YMMV. Helps
@@ -245,7 +243,14 @@
         return item.type.toLowerCase() === predicate;
       },
       tier: function(predicate, item) {
-        return item.tier.toLowerCase() === predicate;
+        const tierMap = {
+          white: 'common',
+          green: 'uncommon',
+          blue: 'rare',
+          purple: 'legendary',
+          yellow: 'exotic'
+        };
+        return item.tier.toLowerCase() === (tierMap[predicate] || predicate);
       },
       // Incomplete will show items that are not fully leveled.
       incomplete: function(predicate, item) {
@@ -328,6 +333,30 @@
 
         return (item.classType === value);
       },
+      glimmer: function(predicate, item) {
+        var boosts = [
+          1043138475, // -black-wax-idol
+          1772853454, // -blue-polyphage
+          3783295803, // -ether-seeds
+          3446457162  // -resupply-codes
+        ];
+        var supplies = [
+          269776572, // -house-banners
+          3632619276, // -silken-codex
+          2904517731, // -axiomatic-beads
+          1932910919 // -network-keys
+        ];
+
+        switch (predicate) {
+        case 'glimmerboost':
+          return boosts.includes(item.hash);
+        case 'glimmersupply':
+          return supplies.includes(item.hash);
+        case 'glimmeritem':
+          return boosts.includes(item.hash) || supplies.includes(item.hash);
+        }
+        return false;
+      },
       stattype: function(predicate, item) {
         return item.stats && _.any(item.stats, function(s) { return s.name.toLowerCase() === predicate && s.value > 0; });
       },
@@ -347,7 +376,8 @@
         return item.name.toLowerCase().indexOf(predicate) >= 0 ||
           // Search perks as well
           (item.talentGrid && _.any(item.talentGrid.nodes, function(node) {
-            return node.name.toLowerCase().indexOf(predicate) >= 0;
+            // Fixed #798 by searching on the description too.
+            return (node.name + ' ' + node.description).toLowerCase().indexOf(predicate) >= 0;
           }));
       },
       light: function(predicate, item) {