+++ app/scripts/store/dimStoreItem.directive.js
@@ -29,9 +29,9 @@
 
 
 
-  StoreItem.$inject = ['dimStoreService', 'ngDialog', 'dimLoadoutService', '$rootScope'];
+  StoreItem.$inject = ['dimItemService', 'dimStoreService', 'ngDialog', 'dimLoadoutService', '$rootScope', 'dimActionQueue'];
 
-  function StoreItem(dimStoreService, ngDialog, dimLoadoutService, $rootScope) {
+  function StoreItem(dimItemService, dimStoreService, ngDialog, dimLoadoutService, $rootScope, dimActionQueue) {
     var otherDialog = null;
 
     return {
@@ -43,7 +43,8 @@
       restrict: 'E',
       scope: {
         store: '=storeData',
-        item: '=itemData'
+        item: '=itemData',
+        shiftClickCallback: '=shiftClickCallback'
       },
       template: [
         '<div ui-draggable="{{ ::vm.draggable }}" id="{{ ::vm.item.index }}" drag-channel="{{ ::vm.dragChannel }}" ',
@@ -59,7 +60,7 @@
         '    <div class="item-xp-bar item-xp-bar-small" ng-if="vm.item.percentComplete && !vm.item.complete">',
         '      <div dim-percent-width="vm.item.percentComplete"></div>',
         '    </div>',
-        '    <div class="img" dim-bungie-image-fallback="::vm.item.icon" ng-click="vm.clicked(vm.item, $event)">',
+        '    <div class="img" dim-bungie-image-fallback="::vm.item.icon" ng-click="vm.clicked(vm.item, $event)" ng-dblclick="vm.doubleClicked(vm.item, $event)">',
         '    <div ng-if="vm.item.quality" class="item-stat item-quality" ng-style="vm.item.quality.min | qualityColor">{{ vm.item.quality.min }}%</div>',
         '    <img class="element" ng-if=":: vm.item.dmg && vm.item.dmg !== \'kinetic\'" ng-src="/images/{{::vm.item.dmg}}.png"/>',
         '    <div ng-if="vm.item.isNew" class="new_overlay_overflow">',
@@ -97,8 +98,20 @@
         });
       }
 
+      vm.doubleClicked = dimActionQueue.wrap(function(item, e) {
+        e.stopPropagation();
+        var active = dimStoreService.getActiveStore();
+
+        dimItemService.moveTo(item, active, item.canBeEquippedBy(active) ? !item.equipped : false);
+      });
+
       vm.clicked = function openPopup(item, e) {
         e.stopPropagation();
+
+        if (vm.shiftClickCallback && e.shiftKey) {
+          vm.shiftClickCallback(item);
+          return;
+        }
 
         dimStoreService.dropNewItem(item);